Monoalphabetic cipher implemented
The idea behind this is to perform cryptanalysis on ciphertext and then
try to predict what letters in the plaintext correspond to certain
letters in the ciphertext.

This is done using frequency analysis of the ciphertext and comparing it
with the English language. This works because in a monoalphabetic
cipher, letter frequencies are maintained.