High-level file for midterm messages
This is a very high level culmination of my crypto code.

It is not very impressive in its ability to decrypt on its own. It is
basically there to run a quick information gathering session on all of
the messages on the midterm and output that information to the console
so I can manually go through and do further cryptanalysis.

It does not currently have polyalphabetic cipher support.