IT WORKS
Assuming a message was enciphered 'kindly' like in homework problems,
this will allow the user to simply run hill_system.py with a --msg and
it will spit out a list of (usually less than 32) potential plaintext
sentences. The user then simply has to scan through the list and look
for ones that are actual English and the encryption/decryption key will
be above it in the console.